Add checks for flag inputs, also renamed the GUI to python_gui.py to avoid confusion
